+++ Tests/LinuxMain.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+import SendBirdSDKTests
+
+var tests = [XCTestCaseEntry]()
+tests += SendBirdSDKTests.allTests()
+XCTMain(tests)